[Refactor] 붕어빵 리스트 컴포넌트 분리 작업
